Fix update logic
