+++ charts/core/src/Chart/hooks/updateUtils.ts
@@ -1,5 +1,3 @@
-import merge from 'lodash.merge';
-
 import { ChartOptions, SeriesOption } from '../Chart.types';
 
 export function addSeries(
@@ -32,9 +30,38 @@ export function removeSeries(
   return updatedOptions;
 }
 
+/**
+ * Method to recursively merge two objects. It should update keys if they
+ * already exist and add them if they don't. However, it shouldn't completely
+ * overwrite a key it's an already existing object.
+ *
+ * They goal is to allow for partial updates to the chart options object.
+ */
+function recursiveMerge(
+  target: { [key: string]: any },
+  source: { [key: string]: any },
+) {
+  const updatedObj = { ...target };
+
+  for (const key in source) {
+    if (
+      typeof source[key] === 'object' &&
+      typeof updatedObj[key] === 'object'
+    ) {
+      // Recursively update nested objects
+      updatedObj[key] = recursiveMerge(updatedObj[key], source[key]);
+    } else {
+      // Update or add the value for the key
+      updatedObj[key] = source[key];
+    }
+  }
+
+  return updatedObj;
+}
+
 export function updateOptions(
   currentOptions: ChartOptions,
   options: Partial<ChartOptions>,
 ): Partial<ChartOptions> {
-  return merge({ ...currentOptions, ...options });
+  return recursiveMerge(currentOptions, options);
 }
