--- charts/core/src/XAxis/config/index.ts
@@ -1,1 +1,0 @@
-export { useXAxisOptions } from './useXAxisOption';
